question_generator_v2/requirements.txt
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
PyYAML==6.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0

